requirements.txt
evds
pandas
matplotlib
python-dotenv
jupyter
pytrends
seaborn
numpy 